p4i integration Integrate 113827 from 3 to main: removing some warnings ... some of them are just been picky part of task#31410
[git-p4: depot-paths = "//depot/qt/main/": change = 113832]

+++ src/kernel/qapplication_win.cpp
@@ -678,7 +678,7 @@ void qt_init( QApplicationPrivate *priv, int )
     HRESULT r;
     r = OleInitialize(0);
     if ( r != S_OK && r != S_FALSE ) {
-	qWarning( "Qt: Could not initialize OLE (error %x)", r );
+	qWarning( "Qt: Could not initialize OLE (error %x)", (unsigned int)r );
     }
 #endif
 
@@ -1801,7 +1801,7 @@ LRESULT CALLBACK QtWndProc( HWND hwnd, UINT message, WPARAM wParam,
 	case WM_GETOBJECT:
 	    {
 		// Ignoring all requests while starting up
-		if ( qApp->startingUp() || !qApp->loopLevel() || lParam != OBJID_CLIENT ) {
+		if ( qApp->startingUp() || !qApp->loopLevel() || (DWORD)lParam != OBJID_CLIENT ) {
 		    result = FALSE;
 		    break;
 		}
